requirements.txt
# Core web framework dependencies
uvicorn
fastapi
starlette
python-multipart

# Template and rendering
jinja2
markdown

# AI and ML dependencies
google-generativeai
numpy
scikit-learn
faiss-cpu
langgraph
langchain-core
langchain-google-genai
langsmith

# Configuration and environment
python-dotenv
pydantic
pydantic-core
typing_extensions

# HTTP and networking
requests
httpx

# Retry and resilience
tenacity

# Database
sqlalchemy
chromadb

# UI Framework
streamlit

# PDF generation
weasyprint
reportlab

# Testing
pytest
pytest-asyncio
pytest-playwright
playwright

# Observability and Monitoring
python-json-logger
prometheus-client

# Utilities
pathlib2
dataclasses-json
protobuf

# Additional dependencies
dependency-injector
beautifulsoup4
docker
flake8
pylint
black
psutil
tiktoken
